CAMEL-14260: Use testcontainers at camel-rabbitmq unit-tests
