+++ components/camel-rabbitmq/src/test/java/org/apache/camel/component/rabbitmq/RabbitMQEndpointTest.java
@@ -119,9 +119,9 @@ public class RabbitMQEndpointTest extends CamelTestSupport {
         customHeaders.put("byteArrayHeader", "foo".getBytes());
         customHeaders.put("longStringHeader", LongStringHelper.asLongString("Some really long string"));
         customHeaders.put("timestampHeader", new Timestamp(4200));
-        customHeaders.put("byteHeader", new Byte((byte)0));
-        customHeaders.put("floatHeader", new Float(42.4242));
-        customHeaders.put("longHeader", new Long(420000000000000000L));
+        customHeaders.put("byteHeader", Byte.valueOf((byte)0));
+        customHeaders.put("floatHeader", Float.valueOf((float) 42.4242));
+        customHeaders.put("longHeader", Long.valueOf(420000000000000000L));
         Mockito.when(properties.getHeaders()).thenReturn(customHeaders);
 
         byte[] body = new byte[20];
@@ -138,9 +138,9 @@ public class RabbitMQEndpointTest extends CamelTestSupport {
         assertArrayEquals("foo".getBytes(), (byte[])exchange.getIn().getHeader("byteArrayHeader"));
         assertEquals("Some really long string", exchange.getIn().getHeader("longStringHeader"));
         assertEquals(new Timestamp(4200), exchange.getIn().getHeader("timestampHeader"));
-        assertEquals(new Byte((byte)0), exchange.getIn().getHeader("byteHeader"));
-        assertEquals(new Float(42.4242), exchange.getIn().getHeader("floatHeader"));
-        assertEquals(new Long(420000000000000000L), exchange.getIn().getHeader("longHeader"));
+        assertEquals(Byte.valueOf((byte)0), exchange.getIn().getHeader("byteHeader"));
+        assertEquals(Float.valueOf((float) 42.4242), exchange.getIn().getHeader("floatHeader"));
+        assertEquals(Long.valueOf(420000000000000000L), exchange.getIn().getHeader("longHeader"));
         assertEquals(body, exchange.getIn().getBody());
     }
