+++ components/camel-rabbitmq/src/test/java/org/apache/camel/component/rabbitmq/integration/AbstractRabbitMQIntTest.java
@@ -14,28 +14,48 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.rabbitmq;
+package org.apache.camel.component.rabbitmq.integration;
 
 import java.io.IOException;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
 import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
-import org.apache.camel.test.junit4.CamelTestSupport;
+import org.apache.camel.test.testcontainers.ContainerAwareTestSupport;
 import org.apache.qpid.server.SystemLauncher;
 import org.apache.qpid.server.model.ConfiguredObject;
 import org.apache.qpid.server.model.SystemConfig;
+import org.testcontainers.containers.GenericContainer;
 
-public abstract class AbstractRabbitMQIntTest extends CamelTestSupport {
+public abstract class AbstractRabbitMQIntTest extends ContainerAwareTestSupport {
     protected static final String INITIAL_CONFIGURATION = "qpid-test-initial-config.json";
     protected static SystemLauncher systemLauncher = new SystemLauncher();
-        
+
+    // Container starts once per test class
+    private static GenericContainer container;
+
+    @Override
+    public boolean isCreateCamelContextPerClass() {
+        return true;
+    }
+
+    protected boolean isStartDocker() {
+        return true;
+    }
+
+    @Override
+    public void setUp() throws Exception {
+        super.setUp();
+    }
+
     /**
      * Helper method for creating a RabbitMQ connection to the test instance of the
      * RabbitMQ server.
+     *
      * @return
      * @throws IOException
      * @throws TimeoutException
@@ -49,7 +69,7 @@ public abstract class AbstractRabbitMQIntTest extends CamelTestSupport {
         factory.setVirtualHost("/");
         return factory.newConnection();
     }
-    
+
     /**
      * Helper method for creating a Qpid Broker-J system configuration for the
      * initiate of the local AMQP server.
@@ -62,5 +82,24 @@ public abstract class AbstractRabbitMQIntTest extends CamelTestSupport {
         attributes.put(SystemConfig.STARTUP_LOGGED_TO_SYSTEM_OUT, false);
 
         return attributes;
+    }
+
+    @Override
+    protected GenericContainer<?> createContainer() {
+        container = isStartDocker() ? DockerTestUtils.rabbitMQContainer() : null;
+        return (GenericContainer<?>) container;
+    }
+
+    @Override
+    protected void cleanupResources() throws Exception {
+        super.cleanupResources();
+
+        if (container != null) {
+            container.stop();
+        }
+    }
+
+    protected long containerShutdownTimeout() {
+        return TimeUnit.MINUTES.toSeconds(1L);
     }
 }
