+++ components/camel-rabbitmq/src/test/java/org/apache/camel/component/rabbitmq/integration/RabbitMQRequeueHandledExceptionIntTest.java
@@ -14,14 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.rabbitmq;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
+package org.apache.camel.component.rabbitmq.integration;
 
 import org.apache.camel.Endpoint;
 import org.apache.camel.EndpointInject;
@@ -29,24 +22,21 @@ import org.apache.camel.Produce;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.component.rabbitmq.RabbitMQConstants;
+import org.apache.camel.test.junit4.CamelTestSupport;
 import org.junit.Test;
 
 /**
- * Integration test to check that RabbitMQ Endpoint is able handle heavy load using multiple producers and
- * consumers
+ * Integration test to confirm REQUEUE header causes message not to be re-queued when an handled exception occurs.
  */
-public class RabbitMQLoadIntTest extends AbstractRabbitMQIntTest {
+public class RabbitMQRequeueHandledExceptionIntTest extends AbstractRabbitMQIntTest {
     public static final String ROUTING_KEY = "rk4";
-    private static final int PRODUCER_COUNT = 10;
-    private static final int CONSUMER_COUNT = 10;
-    private static final int MESSAGE_COUNT = 100;
-    
+
     @Produce("direct:rabbitMQ")
     protected ProducerTemplate directProducer;
 
     @EndpointInject("rabbitmq:localhost:5672/ex4?username=cameltest&password=cameltest"
-                          + "&queue=q4&routingKey=" + ROUTING_KEY + "&threadPoolSize=" + (CONSUMER_COUNT + 5)
-                          + "&concurrentConsumers=" + CONSUMER_COUNT)
+            + "&autoAck=false&queue=q4&routingKey=" + ROUTING_KEY)
     private Endpoint rabbitMQEndpoint;
 
     @EndpointInject("mock:producing")
@@ -66,37 +56,27 @@ public class RabbitMQLoadIntTest extends AbstractRabbitMQIntTest {
                         .log("Sending message")
                         .inOnly(rabbitMQEndpoint)
                         .to(producingMockEndpoint);
+
                 from(rabbitMQEndpoint)
+                        .onException(Exception.class)
+                        .handled(true)
+                        .end()
                         .id("consumingRoute")
                         .log("Receiving message")
-                        .to(consumingMockEndpoint);
+                        .inOnly(consumingMockEndpoint)
+                        .throwException(new Exception("Simulated handled exception"));
             }
         };
     }
 
     @Test
-    public void testSendEndReceive() throws Exception {
-        // Start producers
-        ExecutorService executorService = Executors.newFixedThreadPool(PRODUCER_COUNT);
-        List<Future<?>> futures = new ArrayList<>(PRODUCER_COUNT);
-        for (int i = 0; i < PRODUCER_COUNT; i++) {
-            futures.add(executorService.submit(new Runnable() {
-                @Override
-                public void run() {
-                    for (int i = 0; i < MESSAGE_COUNT; i++) {
-                        directProducer.sendBodyAndHeader("Message #" + i, RabbitMQConstants.ROUTING_KEY,
-                                                         ROUTING_KEY);
-                    }
-                }
-            }));
-        }
-        // Wait for producers to end
-        for (Future<?> future : futures) {
-            future.get(5, TimeUnit.SECONDS);
-        }
-        // Check message count
-        producingMockEndpoint.expectedMessageCount(PRODUCER_COUNT * MESSAGE_COUNT);
-        consumingMockEndpoint.expectedMessageCount(PRODUCER_COUNT * MESSAGE_COUNT);
-        assertMockEndpointsSatisfied(5, TimeUnit.SECONDS);
+    public void testTrueRequeueHeaderWithHandleExceptionNotCausesRequeue() throws Exception {
+        producingMockEndpoint.expectedMessageCount(1);
+        consumingMockEndpoint.setMinimumExpectedMessageCount(1);
+
+        directProducer.sendBodyAndHeader("Hello, World!", RabbitMQConstants.REQUEUE, true);
+
+        producingMockEndpoint.assertIsSatisfied();
+        consumingMockEndpoint.assertIsSatisfied();
     }
 }
