+++ components/camel-rabbitmq/src/test/java/org/apache/camel/component/rabbitmq/integration/RabbitMQRequeueUnhandledExceptionIntTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.rabbitmq;
+package org.apache.camel.component.rabbitmq.integration;
 
 import org.apache.camel.Endpoint;
 import org.apache.camel.EndpointInject;
@@ -22,13 +22,13 @@ import org.apache.camel.Produce;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.test.junit4.CamelTestSupport;
+import org.apache.camel.component.rabbitmq.RabbitMQConstants;
 import org.junit.Test;
 
 /**
- * Integration test to confirm REQUEUE header causes message not to be re-queued when an handled exception occurs.
+ * Integration test to confirm REQUEUE header causes message to be re-queued when an unhandled exception occurs.
  */
-public class RabbitMQRequeueHandledExceptionIntTest extends CamelTestSupport {
+public class RabbitMQRequeueUnhandledExceptionIntTest extends AbstractRabbitMQIntTest {
     public static final String ROUTING_KEY = "rk4";
 
     @Produce("direct:rabbitMQ")
@@ -58,20 +58,20 @@ public class RabbitMQRequeueHandledExceptionIntTest extends CamelTestSupport {
 
                 from(rabbitMQEndpoint)
                         .onException(Exception.class)
-                        .handled(true)
+                        .handled(false)
                         .end()
                         .id("consumingRoute")
                         .log("Receiving message")
                         .inOnly(consumingMockEndpoint)
-                        .throwException(new Exception("Simulated handled exception"));
+                        .throwException(new Exception("Simulated unhandled exception"));
             }
         };
     }
 
     @Test
-    public void testTrueRequeueHeaderWithHandleExceptionNotCausesRequeue() throws Exception {
+    public void testTrueRequeueHeaderWithUnandleExceptionCausesRequeue() throws Exception {
         producingMockEndpoint.expectedMessageCount(1);
-        consumingMockEndpoint.setMinimumExpectedMessageCount(1);
+        consumingMockEndpoint.setMinimumExpectedMessageCount(2);
 
         directProducer.sendBodyAndHeader("Hello, World!", RabbitMQConstants.REQUEUE, true);
